Refactor tracker
